Gil drops :: Beastmen now drop gil. Needs further work.
Detailed: The problem with this is that there is no <player> obtains xx gil. message because the text offset is zone dependant. There also needs to be some kind of field in CMobEntity which says whether this monster drops gil (as some NMs drop gil, e.g. Sky gods but they aren't beastmen) or some other method of identifying whether or not the target monster drops gil.