NMs :: NMs now undergo special enhancements to make them harder than standard monsters of the same level. See details.
The specific calculations are as follows: HP*2.5 STATS*1.5 (STR DEX etc). If the NM is >Lv75 they get an extra 2.5x HP on top of that. Average NM HP at 75 before was 18.2*75^1.2675 = 4331HP. With this update, a Lv75 NM is now 27,000 HP. Sky gods are set to level 100 in the DB (actually a bit too high) and therefore have gone from 11.6k HP -> 73k HP. The real level of sky gods are Lv90 which would be 5441HP -> 34k HP. Of course, for NMs where the HP is known, this should be implemented, but as a rough formula it works out very nicely.