BUG FIXES :: A plethora of fixes have been issued.
[Console]
Console debug print lines have been changed to make it clearer for debugging via the log.

[Ranged Attacks]
The server console will no longer report getAmmoDmg is null. Furthermore, you are now required to have ammo (and it is deducted) when you use a ranged WS.

[Party]
Fixed an issue whereby changing the party leader would make it impossible to see the party member list (incorrectly assigned new party leader).

[Known Issues]
There is a known issue with the inventory whereby the sync in the inventory list is broken between client and server. The cause is under investigation. There are additional print lines in the console to identify this issue. If you find you can't move an item, log off and back on.

When you run out of ammo, the next item you get will appear in your ammo slot. Other that aesthetics, there appears to be no adverse effects to this.