CRASH FIX :: 07-02-2012 crash fixed.
Scenario: Alice tried to /tell Bob who was in the process of logging off. Bob was removed from the zone list (hence not found in the zone) and so a mog house sweep was performed using the map_session list. The thread to kill the entry for Bob had nulled the entry but not yet removed it. The list entry then tried to access the id of the character resulting in the crash.

Cause: The map session character in the list was null.
Resolution: Checks for null in place.