SEARCH SERVER :: Is actually useful for searching now, but hasn't been put through its paces yet!
Details:
The search server can now search via area (or all), level range, job and NAME! Basically most of the 'common' search requirements are covered by this. The 'inv' flag doesn't work yet, nor does region. However, all of these example searches should work:
/sea 50-75 :: Returns all 50-75 people in your current area.
/sea all ar :: Returns all players beginning with "Ar" (case-insensitive)
/sea all ha drk 10-20 :: Returns all players beginning with "Ha" who are between Lv10-20 who are on DRK.

[Fixes and Tweaks]
Console now gives limited information on the search params and results (less spammy).
Added checks for charname NOT NULL, but the underlying cause is still not known.

Please bear in mind it's not practical to test this kind of stuff using 2-3 machines, since you need a lot of test cases for thorough testing. Therefore, the test server (shock horror) will be used for testing this mainly. Whilst it should be stable and I see no reason for it to fall over, it needs to be put through its paces, particularly for larger queries ('all' searches no longer perform COUNT queries so searches take a bit longer and it may be a performance hit with high populations (remains to be seen))