CORE: Added in special spellcasting scripting functions. See below.
There's a new col in mob_pools table for hasSpellScript. If this is 1, when the time comes to use a spell, the server will lookup the name of the monster (same file as onMobDeath etc) and look for the function onMonsterMagicPrepare. In this function, you should return the spell ID you want the monster to cast. An example is included (see /Ru'Aun Gardens/Sprinkler). If you don't want to cast a spell, return 0. 

There may be bugs in this, but I haven't noticed anything yet.