CORE: No longer hard-code which spell types a monster can use.
Added a new col to mob_pools called 'castSpellTypes'. If this column is set to 0, the monster will not cast any spells. This is a 64 bit bitmask which corresponds to what spell types a monster can use. There are 8 bits allocated to each of the 8 elements. See zoneutils.cpp for information on which bit does what. This is used in conjunction with the hasSpellScript flag. If castSpellTypes=0 and hasSpellScript=1 then only the spell script will be used for spells.