SCRIPTS: Added quest "Brygid the Stylist Returns" along with:
An LUA binding canEquipItem(itemid,checklevel) that determines if a given item can be equipped by the player. The second parameter says whether to take player level into account in the check

A convenience function in a new equipment.lua isArtifactArmor(itemid) which will return true if the given item is AF.

CORE: Moved the MAX_ITEMID, MAX_DROPID, and MAX_LOOTID from the cpp file to the header so they can be used from other places.