Implement PlayTime function and fix tractor
PlayTime function solved the bug of warp to homepoint on death. The
problem is in the packet zone_in, when the pj death, send this packet
with playtime set to 0, the client doesn't understand that a player
whith a platime of 0 has 0 hp and warp to homepoint.

Tractor now working after fixed the bug of warp to homepoint at death

I've tested on my private test server.