SCRIPTS: Guild Image Support Added. Fixed an issue with signing up for the guild @ full inventory. CORE: Prepare for EFFECTFLAG_SYNTH_SUPPORT. SQL: Fix a typo spawn coordinate in my last commit.
Please note: I got all of but 2 NPCs to work: Azima and Fatima. Also keep in mind that the 0x31 packet is not being handled at the moment. Do NOT pick the cutscene option to see recipes, else you will get frozen in the cutscene until its properly implemented in full.