This Yagudo Sentinel was spawned in an invalid position.
It was spawned mid air, client renders the model at ground level on the floor below causing players to wonder why the can walk right up to it yet be told they are too far away to attack.

This change moves it to the ledge next to its mid air position.