Made a semi educated attempt to implement Cura/Cure II
Complete with Afflatus Misery bonus. Primarily used BGWiki
as a source:

http://wiki.bluegartr.com/bg/Cura
http://wiki.bluegartr.com/bg/Cura_II
http://wiki.bluegartr.com/bg/Cure_Formula
http://wiki.bluegartr.com/bg/Cure_Potency

I googled around on this a lot and there does not seem to be a lot of concrete info, so I made it as close to within the known guidelines as possible. Even if its not perfect, it is workable and should be reasonably close to what is expected. Additional notes in the comments.

Also fixed up the SQL for these spells. The targetLevel was wrong as these are self target only. I also updated the fixed enmity values to match BGWiki.