April Update: Fix Promyvion-Dem & Promyvion-Mea Post-Patch
The LUA scripts for warping with the Memory Recepticles needed to be fixed in these two zones. In spire of Dem, the NPC Ids post AchieveMaster were off by one. This caused you to be unable to enter the BCNM. Promyvion Holla had no issues.

Ran all 3 promys tonight. With the fixes in this commit, all three work again.