CORE: Implement Remaining Missing NIN Spells/Abilities/Traits
This changes attempts to implement all remaining missing features for NIN, up to Lv99

The following changes are implemented:

SPELLS:

Aisha: Ichi - Lowers the targets attack by 15%. Corresponding tools / toolbags / and scrolls have been finished as necessary as well.

http://wiki.bluegartr.com/bg/Aisha:_Ichi

Yurin: Ichi - Gives the target the effect Inhibit TP, which lowers TP gain. Lowers TP gain by 10%. Corresponding tools / toolbags / and scrolls have been finished as necessary as well.

http://wiki.bluegartr.com/bg/Yurin:_Ichi

Myoshu: Ichi - Grants the caster the effect of Subtle Blow Plus, with a Subtle Blow Bonus of +10. Corresponding tools / toolbags / and scrolls have been finished as necessary as well.

http://wiki.bluegartr.com/bg/Myoshu:_Ichi

Migawari: Ichi - Makes the caster immune to "severely damaging attacks" for the next minute. These are defines as attacks that do damage within a certain threshold of the caster's max hp. From best available info, the threshold starts at 100% of the users max hp, and is lowered by 1% per 5 ninjutsu skill. Corresponding tools / toolbags / and scrolls have been finished as necessary as well.

http://wiki.bluegartr.com/bg/Migawari:_Ichi

Kakka: Ichi - Grants the caster the effect of Store TP, with a Store TP bonus of +10. Corresponding tools / toolbags / and scrolls have been finished as necessary as well.

TODO: Ryuno base price does not seem to be listed on any of the major wikis. If someone with retail is willing to check this for me I will happily put it in the SQL.

http://wiki.bluegartr.com/bg/Kakka:_Ichi

ABILITIES:

Issekigan - NIN 95 - Substantially increases parrying rate for the next minute. The exact rate is not know, but from the testing linked to from the BGWiki article, it is thought that if you have already capped out your parry rate at 25%, Issekigan will boost you beyond that by another 25% to cap out at around 50% for the duration of the effect. The user will also be granted an emnity bonus for each successful parry.

http://wiki.bluegartr.com/bg/Issekigan

Mikage - NIN 96 - NIN One Hour Ability - Makes your main hand attacks vary with the number of utsusemi shadows that are currently up.

TODO: Determine correct animation. I can't find any screenshots or videos of it, and it is not listed in Altana Viewer. If anyone is willing to help me out with that it would be appreciated :p right now it has the same animation as Issekigan set.

http://wiki.bluegartr.com/bg/Mikage

TRAITS:

Tactical Parry - Grants a TP bonus on each successful parry. Implemented for all job / tier levels that have it. Conflicting information between FFXIcyclopedia / BGWiki as to whether the trait is affected by Store TP or not. BGWiki says no, and seems to be more reliable on this fact, as FFXIclopedia's info is listed as unconfirmed, so the implementation is not affected by Store TP.

http://wiki.ffxiclopedia.org/wiki/Tactical_Parry
http://wiki.bluegartr.com/bg/Tactical_Parry

Mag. Burst Bonus - Grants a bonus to the magic burst damage.

http://wiki.ffxiclopedia.org/wiki/Magic_Burst_Bonus

Skillchain Bonus - Grants a bonus to skillchain damage.

http://wiki.ffxiclopedia.org/wiki/Skillchain_Bonus

This collectively should make nin fully functional all the way to 99.