CORE/SCRIPTS: Refactor Migawari Entry Point
Push the effect out of the calls across the core code. Replaced with an intermediary helper function. The differences between Migawari and similar spells are as follows: the threshold and reduction percent (stored in the effect because they don't always stack together), and whether or not the effect is removed after one hit (the bool flag).

The previous generic method was already equipped to handle that, except, by doing so calls for other spells would be duplicated throughout the core code. The new intermediary method will solve this and serve as the extension point for similar spells, giving us the least amount of copy/paste.