May Update: Text ID Updates
Used the batch job I developed in #215, plus a pass of manual review. These should be good.

As I manually fixed things that could not be auto-fixed (things polluted with bad characters, for example), I tried to update the comments to be more automation friendly for the job next time.

The new batch job has an option to write filtered version of the dialog tables where it has attempted to strip out all the bad characters. I used these as a basis for selecting a new comment - trying to strike a balance between having a meaningful comment, but still allowing the batch job able to find it after stripping out bad characters.

With these updates, the results are reproducible using the job with no manual edits; if you run another pass of the job, it agrees with the results original results with no manual reviews flagged, proving that the updated comments work. This should leave us in good shape for automatic future updates.