Implement Thief Ambush merit ability
Attempting to implement a check for the Thief Ambush merit.  Each merit point is supposed to increase the attacker's accuracy by 3. First it will check to see if the attacker is a PC, then it will check to see if that PC has the TRAIT_AMBUSH merit.  The value retrieved from the TRAIT_AMBUSH merit will be added to the offsetAccuracy variable.  The increased accuracy will then be used in the normal calculations that follow.  This seemed like the right place to implement this trait.  Wanted to submit the suggestion to those that are more familiar with how this might work.