June Update: NPC Id Updates
The output of this commit was semi-automatically generated with the
NPC Id Update batch job.

There does not seem to be many changes this time around, and some of
the ones that are look like they could be things from a previous update
that the earlier version of the batch job was not smart enough to detect.

NOTE: This job still struggles a bit due to mass inconsistencies in
our naming convention for some NPCs. I apologize in advance if anything
is incorrect.

Additionally, due to the inconsistencies in naming, this job has to make
use of a mapping file that compares the state of the ids
from before the version update, to after against the state of POLUtils
before/after. As such, for NPCs that must be mapped (ones with non-unique names,
or formatted significantly different than POLUtils) the job is only as good
as the mapping file - any of those NPCs that were broken before the update
will not be fixed by the job and must be fixed manually as they are encountered.