CORE: Add Ability to Cap Normal Monster Level Ranges
As we start to code support for playing Level 75+, one thing that needs to be addressed is a way for the new exp mobs not to ruin the play experience of those wishing to play classic FFXI.

Currently these mobs, with the exception of the crabs in Vunkeral Inlet, are all set to 81-84 in the database. That level is low enough that they wont destroy level 75 players, but too low to level to 99 on.

This commit introduces new settings.lua variables to cap the level range for normal zone mobs. The defaults are set to 81 & 84 respectively, to maintain current behavior with the default level cap of 75

This will allow us to start putting retail-correct levels in the database, without destroying the play experience for Level 75 players. This will also start to make Level 99 play more playable

The next commit will start to adjust level ranges for the mobs based on the recorded ranges in the various wikis