Update feral_howl.lua
applied changes related to properly calculating resistance and duration based on what the server should return from the MOD values. Added better description of what the script addresses in the introduction section to include references. The math was a lot easier when the assumed return values were 1 for feral howl merits instead of 5.