NPCID - SOA Zones - POLUtils Names Updates
Now that gedads finished his updates in these zones, I have scraped the POLUtils names for these zones. I made a couple of corrections based on the results, it looked like there was one area where a duplicated row was throwing off the ids by 1 still.

More preparation for automation with the next generation of the NPC ID batch job.