NPCs: Introduce Expansion Restriction Support
This commit introduces support for restricting npcs by expansion, using the same approach that was used for spells/abilities/traits

This gets rid of npcs that are not relevant to the era that the server wants to play in - ex: Cavernous Maws without WOTG, Geo stuff from SOA, etc

The approach for tagging npcs is the same as for the other files - the expansion tagged is the latest expansion out at the time the npc was introduced

I only hit obvious ones and tagged expansion zones with their expansion - I am sure they are more. If there are specific npcs that should be mass tagged with a specific expansion, I have a script that can do this by polutils name (will commit it in my next batch update, coming soon)

This makes things more accurate to the era a server wants to play in, as well as reduces memory and resource waste for those servers by not loading unwanted npcs.

I have tested disabling all expansions and re-enabling and the functionality appears to work. I don't get any startup warnings with all expansions disabled. If you find any issues though feel free to tag me.

-Add RESTRICT_BY_EXPANSION setting. Toggle this off if you don't want the functionality

-Fix memory leak on Expansion Enabled Check
    *Use std::string a suggested by @teschnei
    *Cache results in a std::map to avoid hitting the lua file every time