Lua Function TellToPlayer(), GM Command @setpoint
TellToPlayer() is used to display messages to the player without the
message "----== SystemMessage ==----" preceding each line.

@setpoint allows the user to save up to 9 different locations with a
short (44 char) description and to selectively 'warp' to them as needed.
I use this when testing scripts involving several different locations.
The @setpoint command makes use of the function above, TellToPlayer(),
to display usage instructions and a small table listing the descriptions
of the currently saved set points.