Updates to Kirin Changes
Kirin now uses the Astral Flow ability.  Spawning of the pet has been
deferred to a new mob_skill.
Two slight differences still exist between this and retail.  The avatar
has a casting time instead of being instant.  This leaves open a chance
the target can outrun the ability (and potentially, continue to outrun
it).  I could not find a way to remove the casting time, so I'll take
any suggestions if there are any.