SCRIPTS: Update Sobane.lua
As was written, you were unable to trade a Kindred Seal to this NPC and complete Limit Break 4 if you had either "Signed in Blood" or "Flyers for Regine" quests active because of the "elseif" structure. Changed the statements to all be "if" so these trade actions will act independently.

Update Sobane.lua