Scripts: Fix zone to Stellar Fulcrum from Upper Delkfutt's Tower
CS IDs 0 and 1 are for the two porters, while CS ID 2 is for the door at the very bottom of the zone. The script was using 1 for the porter to Middle Delkfutt's Tower, and 2 for the porter to Stellar Fulcrum, and 0 was being unused. So I simply bumped down the event numbers by 1. Now the porter at the top doesn't use the CS of an opening door either.