Scripts/SQL/Core: Implement Angon + item depletion for Repair/Tomahawk.
Scripted Angon, inflicts 20% defense down, requires Angon item to be equipped.
Changed the merit power of Angon to 15 for the duration (15 + merit power for 30-90 sec depending on investment).
Gave the proper model for the Angon item, as well as the Throwing Tomahawk.
Made the Angon/Throwing Tomahawk/Automaton Oils deplete upon using their respective ability (Tomahawk ability is currently unscripted so I merely put in the framework for it - whoever scripts it has to make sure to have the ammo ID check in OnAbilityCheck).