Scripts: Fix NPC for Windurst 7-2
The previous NPC in the mission, Kerutoto sets the missionStatus var to 2. Currently, players who start the mission are unable to complete the mission because it's checking for the var to be something it can't be set to without GM help.

https://github.com/DarkstarProject/darkstar/blob/master/scripts/zones/Windurst_Waters/npcs/Kerutoto.lua#L260

That's the only instance of the previous NPC setting the missionStatus var while...

https://github.com/DarkstarProject/darkstar/blob/master/scripts/zones/Windurst_Waters/npcs/Kerutoto.lua#L68

That is the event is played. At the end, it's set to 2, and then there's another event that merely changes the text.