Scripts: Animsub resets on respawn
These mobs are supposed to have their animsubs reset on spawn because they'll have different ones upon death. Also, the SetMobSkillAttack thing doesn't reset on death, so it needs to be set to false or else it will use the flight mobskill attack.

Note that Aerns should be using Animsub 1 as the default because going from 0 to 2 (bracers) makes them disappear for a second, while they don't when going from 1 to 2.