Scripts: Correct weather check for Qu'Hau spring
After taking a look in zone_weather.sql, I noticed that Ro'Maeve can't have clear weather, at all. I'm assuming the data is correct, particularly because the image on the wiki shows the filled spring and the reflection of the sky is partly cloudy, which is what happens during the sunshine weather.

Also I did bad math in my previous fix to this spring and gave it an extra game hour, so I'm bring it back down.