Scripts: Setting for Dynamis entry to count by midnights
If the setting is enabled, the entry time is simply set to the previous server midnight instead of current time. This will allow for the next entry time to be the number of midnights equal to the setting for days for next entry (ie. if it's set to 3 days for next entry, it will go by 3 server midnights instead of a full 72 hours).