Scripts/SQL: Fix the spawning of TEs in Dynamis - Xarcabard
The checks were being thrown out of whack because of onMobDeath running for every player in the group, so I borrowed code from the Maat's Cap check to fix it. Each trigger mob will set a specific bit on the appropriate variable if it's not already set, and when the variable has reached the correct value (ie. all triggers are dead), the TE will spawn and variable is cleared.

I also corrected the mob group of two TEs, as they were spawning as eyes instead of statues.