Scripts/SQL: Add/update mobskills for (proto) Omega/Ultima.lua
-Corrected the animations for Omega's moves
-Did some scripting for all four mobs, mostly Proto Omega and some Proto Ultima.
-Added Particle Shield, Ion Efflux, and Discharger.
-Updated the checks on the other mobskills due to the scripting done for Proto Omega/Ultima.
-Separated the mob skill lists for the four mobs, as they all have different moves.
-Removed Equalizer due to #2450

Citadel Buster was not added because I couldn't think of a good way to have it triggered. I would think that SE uses a dummy mobskill to trigger the countdown start, but I have no idea what they do for sure. Nuclear Waste doesn't have a good way to be added either, so the six elemental moves will remain usable at will during the time they're supposed to be usable.