Scripts: Remove hardcoded itemid check on Relic WSs
Since the only way to get access to them is to have a completed relic/stage 4 weapon in Dynamis, there's no point in having a check for Aftermath. Whoever decides to add those level 85 latent weapons will have to add a specific check that that point in time.

Other fixes include:
-Refinement of Aftermath application (removed the if/elseif trees)
-Corrected the effect inflicted to the enemy from Randgrith (from weight to evasion down)
-Corrected the Subtle Blow values for Final Heaven and Blade: Metsu Aftermath.
-Added Attack Down status on Gate of Tartarus.