Core/SQL: Implement Song Recast Delay
Reduction takes place at the end after all multipliers so it would be a reduction in seconds, as per retail. https://www.bg-wiki.com/bg/Song_Recast_Delay

Note that due to the value for merits being stored as a uint8, I am multiplying the value by 1000 in the core instead of setting the value in merits.sql to 1000.