Scripts/SQL: Fix some mob spawn point locations
For the one THF AF NM, I made it set the spawn point before spawning it, instead of attempting to move a mob that's in combat. Since the function that was used before explicitly has a comment saying not to use it on mobs in combat. For Armed Gears, I just set the spawn point to the pos of the ??? that spawns it, which is a hell of a lot more correct than spawning outside of the room and in the floor.