Core/Scripts/SQL: Fix Ghrah aggro
Created a binding to set the aggro instead of using the one to set true detection (ie. what it should have been to begin with). Also made it set aggro in onMobFight as well since it does change form.

Also changed the detection for Jailer of Fortitude to be sound like the rest of the Ghrah instead of sight, and removed the declaration of a variable that never gets used (I forgot to remove that forever ago).