Fixed issues with the quest 'Mom, the Adventurer?'
Fixed some issues with the quest 'Mom, the Adventurer?'. After reaching
Bastok rep 2 or higher the player was able to start the next quest
without completing this one. In addition, even if this quest had been
accepted, the player would be unable to complete it after reaching rep
2+. A check was also added to fix a player who has already had this
quest bugged for them so they should now be able to complete it.