SCRIPTS/SQL: TOAU-15 The Black Coffin
Instanced battle on the Ashu Talif. There are two waves
in this fight. An initial wave of 5, then a wave of 4
plus the captain.

The captain is initially untargetable and up on high.
After the crew is engaged, he will jump down with a
special animation and join the battle.

When the captain has <=20% HP left, he will give up.
Everyone will disengage and the instance will end in
a cutscene.

Gessho will assist the player in battle. If the players
stalls for too long, he will attack on his own, otherwise
he will assist the players when they engage. Gessho is
allowed to die in this battle and it does not constitute
a failure of the mission.

Sources:
http://ffxiclopedia.wikia.com/wiki/The_Black_Coffin
https://www.bg-wiki.com/bg/Aht_Urhgan_Mission_15
gedads data: http://pastebin.com/w0TqRPMy