SCRIPTS: TOAU-25 Playing the Part
Cutscene at Salaheem's Sentinels

Note: there was conflicting information as to whether this was
triggered on the Salaheem's Sentinels region, or by talking to
Naja directly. The cutscene refuses to trigger if initiated
by talking to Naga, but will trigger if initiated on region,
so I'm assuming region is correct.

Sources:

http://ffxiclopedia.wikia.com/wiki/Playing_the_Part
https://www.bg-wiki.com/bg/Aht_Urhgan_Mission_25
gedads data: http://pastebin.com/w0TqRPMy