SCRIPTS/SQL: TOAU-31 Shades of Vengeance
Instanced fight in Periqia. Requires a special assault permit
key item. The player receives one to start with from the previous
missions. Replacements can be obtained once per Vana'diel day
at the staging point if the player loses the instance. Party
members do not need a key item to assist.

The fight is against 10 lamia. BGWiki says a random amount must
be defeated. FFXIcyclopedia says 3, but that may have been that
person's experience. In this implementation, a random amount
of at least 3 is required to complete the instance.

The following things need to be confirmed to improve this fight

-Instance ID:
This is most assuredly wrong, but it wasn't on the list. Picked
one out of the way. The menu comes up right and it works, but
a message is printed to the player saying they are doing
Better Than One when they zone in. This does not effect the
mission. Playing with the CS params & instance id did not make
a difference on this. I'm not sure why this is.

-Starting position. Made up a position based off of watching some
youtube clips. Works for now, and easily fixed in the SQL when we
have a real one

Sources:

http://ffxiclopedia.wikia.com/wiki/Shades_of_Vengeance
https://www.bg-wiki.com/bg/Aht_Urhgan_Mission_31
gedads data: http://pastebin.com/w0TqRPMy