SCRIPTS: TOAU-46 Imperial Coronation
Cutscene at the Imperial Whitegate. You must equip appropriate
armor and unequip weapons, as in Guests of the Empire.

You will be offered a choice of 3 rings. If you do not choose now,
you can come back to the gate and choose later. You also receive
an Imperial Standard furnishing. You can also get this from the
gate later if your inventory was full.

The time the ring was received is tracked, but the NPC to rechoose
rings (Nadeey) is not implemented because I don't know the right
event.

Sources:

http://ffxiclopedia.wikia.com/wiki/Imperial_Coronation
http://ffxiclopedia.wikia.com/wiki/Aht_Urhgan_Mission_47:_The_Empress_Crowned
gedads data: http://pastebin.com/w0TqRPMy