Trust magic is now properly added to the trust menu instead of the magic menu.
The reason these crashed before was their group was invalid and the client knows where to put them on its own and doesn't like to if you try and tell it the wrong thing about where these belong. For this reason, these are not commented out - however they will simply do nothing until someone does the needed core+script work to make them "real".

A dummy script is included to show that the script is actually run.