Move Field Manuals, Grounds Tomes, and Voidwatch NPCs away form needing static IDs in zone scripts..
Note: npc_list.sql looks more changed than it is. Due to formatting and re-dumping parts of it, I broke git blame. But if an NPC is wrong it was already wrong, I wasn't there to change NPCs just re-tag and re-flag.

btw, THOU SHALT NOT QUOTE NON STRINGS. Cough.