CORE: Handle potential compression failures
Halves amounts of packets to send each time compression fails.
If we run out of packets, we remove 1 and try again.
If we can't compress anything we bail out.

I'm not sure if the compression fails anymore with the output buffer
size increased. This commit may be removed if we never get warnings from
zlib.cpp anymore.