Fix item mods and other stuff for items 10327-10450: Now with Chocobiscuits.
Also, removed ramblers cloak. Nothing was scripted there. It was originally scripted for the item latent, but the item has the proper latent...

And, last, can someone check my hipster set logic in gear_sets.lua, line 261-281, for Phorcys Armor Set? I modeled it after the first hipster set, but not completely sure if it is the best way to do it. I will test this evening.

Test failed, still unsure what the issue is.