Progression fix for Trouble at the Sluice w/DRG quest flagged
Was not able to complete "Trouble at the Sluice" if you talked to Arminibit who starts you on the path to unlock DRG ("The Holy Crest"). This fixes that so you can progress to finish "Trouble at the Sluice" and then refers you back to the DRG quest if you still want to unlock it.