Implemented BST ability "Tame"
So, not much for reference online. I got most of my info from:
http://ffxiclopedia.wikia.com/wiki/Tame
and made 'best guess' as far as the values and percentages based on what is already in dsp and from what I "remember". Please let me know if there's anything I can/should do to make this better.