Restrict character creation to alphabetic only, like retail. Non standard "vanity" names can still be set by editing the database.
Also protects DB from any possibility of Little Bobby Tables joining the server. I know it seems silly to both restrict to A-z in one part then escape the string as well, but I did not want to duplicate the block that sends the error message to the client. If we knew more of the error messages (we need some captures there?) we could put that into a function to call instead.

Fixes #4502