quest(a minstrel in despair): Allow obtaining additional poetic parchments
If you dropped your poetic parchment during "The Old Monument", the player var from the quest would prevent you from obtaining additional parchments which you need to complete the following quest. I've delayed the player var reset to the completion of "A Minstrel in Despair".

Regardless of the number of poetic parchments the player receives, he will only be able to complete "A Minstrel in Despair" once.