Minor updates for qm8 (Lamian Key)
npcutils works here but still does not give retail message. 
npcutils ALSO complets the if statement even though the condition is not met (key not given) and causes the timer to reset so that you cannot get another key until the next day. This is NOT retail behavior.
the way it is scripted now is retail accurate with messages as well as timer resets.
i don't see any reason to script it two different ways, one for dsp and one for personal use to maintain retail accuracy just to say we are using the shiny new tools.