helm(ifrit's cauldron): Remove bad mining points
This point was removed: { -16.149,  26.090,  137.087} (Map 8? H-5)

I could not find it listed either on Vana'diel Atlas, which wiki, bg, and others use as a primary source, or on any jp sources, e.g. http://ff11db.sakura.ne.jp/database/syuusyuu/saikutu.html#ifrit. Note that this actually corresponds to a map not properly listed by Atlas. I'm not sure why Atlas consolidated them.

This point was also removed: { -10.390, 13.376, 92.300} (Map 7 H-5)

It is basically a duplicate of { -3.945, 9.675, 95.107} (Map 7 H-5), which is just a few yalms away.