Updated Citipati respawn timers
In advance, please forgive me for this "well actually herpderp" correction.

Citi repops every 3-6 game nights, rather than a fixed timer on Earth hours. This means he can be killed as late at 04:00 on Night 1, and respawn as early as 20:00 on Night 4 (assuming no PHs for this math). One Vana day is exactly 57m 36s so the minimum respawn is ((57*60)+36)*(2+(16/24)), or 9,216 seconds. Add three full Vana days for the minimum time on the longest window.