Re-sequence mob_groups by new primary key, add name column for human readability.
Table has been re-dumped which has lost some useful commenting (example: which floor of promyvion a mob group is, because there are multiple lv ranges of the same mob in the same zone). These can either be restored at a later date or more preferably, just edit the mobname for that row - the name column of mob_groups is not currently used anywhere so no core code had to change to accommodate this nor does it actually matter if the name is 100% accurate to what is seen in game.