Refactor npc IDs to zone-specific IDs.lua files
- The peddlestox keys in `zoneData` have been dropped. Peddlestox's id now
lives in the zone-specific IDs.lua files, and is passed in by Peddlestox's
NPC scripts. Peddlestox's `day` key is now its own key `zoneData`

- The digsites keys in `zoneData` have been dropped. Their IDs are now stored
in the zone-specific IDs.lua files, and are passed in by the NPC scripts of
their corresponding QM's.

- Dig site position data is no longer stored, and is instead accessed via
`entity:getPos()`

- Dig site QM number is no longer used as an index, and dig site "chest"
number is no longer stored explicitly. Instead, chest number is derived from
the indices of `npc.BEASTMEN_TREASURE` in `IDs.lua`, and QM number is noted
in the comments. Digsites can now be expressed as a number 1-8, and their
"chest" argument is the same number -1.

- Weighted pool assignments now gets their zone id indices from the global
zone table