Added Lycopodium NPCs and their teleports
Added logic for Lycopodiums in past and for blank points in present
that allow teleport after trading specific flowers.

Fixed spawn point in npc_list.sql, there was one posted as apocalypse
but there was no script for it as it was an unnamed npc without logic.
The correct one is named relic and it already has a valid relic.lua and
valid location in sql file.

Submitting this as a draft to go through the trial by fire, once IBM
merges his change, I will update it with the new charvar function
name he changed it to.

(Left NPC name field blank as that's their valid name, wasn't sure
if I shoulda just deleted the field but left it empty so it matches the
name-in-game.