Refactor the order of operations for how damage is applied from spells.
This primarily refactors the order of operations such that any magic
spell that deals damage now handles the effects from the damage
immediately after.

This is necessary as a prerequisite for implementing spells like
Pinecone Bomb, which reapply a sleep effect after the damage has been
applied.

The prior implementation would apply the damage and then apply the sleep
all in the Lua code before returning back to
CBattleEntity::OnCastFinished(), which would then remove the sleep since
damage was taken.

After this refactor the damage and debuff behavior from the damage all
occur on the Lua side. That way Lua scripts for spells can reapply
sleeps and have it function properly.

This was modeled after how TakeWeaponskillDamage() was implemented to
achieve a similar order of operations.

NOTE: I tested manually by sleeping mobs and applying the various types
of spells to verify that sleep is canceled correctly. Also tested a
pinecone_bomb script (not included in this PR) to verify that
re-applying sleep would work.

I also audited every Lua script that includes a spell that deals damage
to cover all possible code paths.